requirements.txt
PyYAML>=6.0
requests>=2.0
apscheduler>=3.9
python-dotenv>=1.0
pytest>=7.0
